Crate hyperion_arch

Modules

context

cpu

int

paging

pmm

stack

syscall

tls

vmm

PageMap is the Page Table Manager

Constants

IA32_TSC_AUX

Functions

cpu_count

cpu_id

dbg_cpu

done

init_bsp_cpu

should be called only once and only by the bootstrap processor before cpu_id

init_smp_cpu

should be called only once per cpu and before cpu_id

reset_cpu_id⚠

Safety

rng_seed

set_cpu_id⚠

Safety

spin_loop

stack_pages